Extract normalizeNotFoundUrl helper
